Refactor graph executor
